Moved get_resolved_parameter_values_from_env_name to config module
